Build readme and program to decode F-BIDR data
The readme is a definite work in progress. I started going over the
F-BIDR data in the readme in length. I decided that this was the wrong
place to do it. I'd be better off explaining in that level of detail
through a program. So I'm creating a way to describe binary data
declaratively (emphasis on what it is and not how to read it, or rather
the how to read it comes straight out of the what it is).

Will start removing most of the explanations of F-BIDR format from the
README, and perhaps use python to generate some better diagrams later
based on the binary data thing I'm doing.

Changes to be committed:
      modified:   .gitignore
      modified:   README.md
      new file:   attrs_structs.py