Add notes on creating declarative binary format
Writing this is a little trickier than I thought. But I think it beats
out writing something finicky as fuck. The hope is that, in the end,
this can serve equally well as documentation as it does code for reading
the F-BIDR files. This way it'll serve someone well even if they don't
use the program.

Changes to be committed:
      modified:   attrs_structs.py
      new file:   notes-on-declarative-binary.md