Fix logic with parent dicts for Series
What remains: all of the compounding record functions need access to the
records created thus far. If records need it to refer to an existing
member, and to pass it to contained series records.

I could write logic in each composition record to recieve a root
dictionary of interpreted records and figure out what the current record
they're looking at needs, and pass that information into the current
record, but that sounds inconvenient. I'd rather give all the compound
records the info that any other compound record may need, and allow each
compound record to take what it needs from the record containing it.
This way what a record needs is written once. Not many times, in both a
parameter list of the __call__ method, and in the parent record when it
figures out what to pass to the current method its processing.

Changes to be committed:
      modified:   attrs_structs.py