Finish Series, List, If records, update tests
They finally all work as they should, though perhaps a little clunky. I
tried first to write a function that iteratively gives paths from the
root to any node in a tree. That filled in the blanks for trying to
iteratively build a tree. I replaced the path (location info) with the
node to use in the new tree (another way of specifying location). It
worked, and was still a little confusing, but still worked.

Must profile.

Changes to be committed:
      modified:   attrs_structs.py
      modified:   notes-on-declarative-binary.md
      modified:   records_tests.py