Fix floats again, add float tests
An explanation of what went wrong, and how the fix happened.

I took a look at libvaxdata's source (https://pubs.usgs.gov/of/2005/1424/libvaxdata.tar.gz)
and noticed that they had a tests file named tests.c. It had binary vax
floats and the floating point value this was supposed to represent.
Bingo, I could now make test cases. Assuming they were correct, I saw
the following:

1.0 represented by : 0x80400000
-1.0 represented by: 0x80C00000

The only difference between the two is the sign bit, which means this
tells me where the sign bit is.

    8040 0000 : 1000 0000 0100 0000 0000 0000 0000 0000
    80C0 0000 : 1000 0000 1100 0000 0000 0000 0000 0000
                        -
                here is the sign bit

That means bit 15 is the start of the 2nd byte. That suggests the bit
order was really:

    7 6 5 4 3 2 1 0 15 14 13 12 11 10 9 8

as opposed to

    0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15

This "bit order" business doesn't change which bits were where, of
course. Just which bit was meant by "the ith bit". I figured that bit 7
would've meant the 7th bit from the right, treating the 4 bytes as a
stream of bits (latter order). Apparently it means "within the 1st byte,
the 7th most significant bit" (former order).

So, the bits within a byte were numbered according to significance, in
each byte, as if each one were to be taken as a binary number. That
makes no sense to me whatsoever, but that seems to be the case.

My code was written under the assumption of the latter bit order as
opposed to the former. So I changed _byte_to_bits to give bits from a
byte in the reverse order. Thus, within each byte, the 1st bit given was
the least significant and the last bit given was the most significant
(the opposite bit order of how one would write down a binary number).
This changes the bit order from the former to the latter.

Lastly, I did the creation of the fraction wrong. I copied straight a
document that described the mantissa as a binary decimal, so I wrote

    return 0.1 + accum

0.1 as a binary decimal is 1/2, but in python that's a decimal decimal,
which means 1/10.