Attempt two: change pixel overwrite rules
Originally pixels from later records overwrite pixels from prior
records. But that means the shit pixels at the edge overwrite the
potentially decent pixels from the middle of another orbit.

Invalid pixels are all 0 (though some valid pixels may be 0, though it
is rare). So I changed the rule so that if a new pixel is 0 and the old
pixel is non-zero, the old pixel stays. Otherwise the new pixel
overwrites the old one.

Another overwrite rule could be taking the maximum of the two pixels.
Might be faster. There's still some black speckle introduced by the
merging of orbits, and I know that black speckle is at the edge of one
of the orbits; if the speckle survived, that means it's not full black.

Yep, just did a test with GIMP: those speckle pxes aren't pure black,
they're just close.

Another overwrite rule could deal with valid/invalid pixels. I could
maybe create that with an array mask or something. Could take more time
to read and process for an image, though.