Add support for a draft mode
To enable draft-only content+effects, set environment variable POLLEN to
"draft". Example:

    POLLEN="draft" raco pollen start

or

    POLLEN="draft" raco pollen render {file}

Either would work. Draft tag splices content into a document only when
draft mode is on. The markdown support isn't great yet for draft, there
are some bugs. I will fix some, but not all. In particular, I don't care
if I can't get markdown content from inside a draft tag to play nice
with content outside of it. It should be used for large sections, not
for small inline pieces.

It's for work that is ongoing and will need quite a bit of time to be
finished. An alternative would be making a different branch. I could
have draft work on there just fine. Though some visual distinctions do
help.

It may be better to reduce the complexity of the draft effects and start
a new branch for ongoing work which will not get done soon.

Changes to be committed:
      modified:   docs/style.css
      new file:   docs/style.css.pp
      modified:   pollen.rkt