Change image data in f_bidr to be numpy array
The change may result in overall speed gains in reading image data.

Without creating a mask for the data, we get a very large speed gain of
around .6s on my machine (that's around 2/3 as long as before)

With creating a mask, we take about 1.2s more, which is poor
performance. So there is now a boolean to control whether masks are
made. I'm sure it will be overall faster for creating images to create
the mask all at once, but would waste the speed gains for when masks are
not needed.

Changes to be committed:
      modified:   f_bidr.py